pyproject.toml
[project]
name = "lite-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.73.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
